# Fix scheduler perf dropping after 128 seconds